requirements.txt
alpaca-trade-api
pandas
numpy
scikit-learn==1.3.2
joblib
pytest
python-dotenv